requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
tqdm>=4.66.0
pydantic>=2.6.0
